+++ packages/ui/src/index.tsx
@@ -1,4 +1,6 @@
 export * from './components/brand/app-title';
 export * from './components/brand/splash-screen';
 
+export * from './components/form/select-field';
+
 export * from './utils/class-names';