+++ apps/backyard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # backyard
 
+## 0.0.9
+
+### Patch Changes
+
+- Create a new streamlined multi-purpose select-field
+- Updated dependencies
+  - ui@0.0.4
+
 ## 0.0.8
 
 ### Patch Changes