+++ apps/foh/CHANGELOG.md
@@ -1,5 +1,12 @@
 # foh
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - ui@0.0.4
+
 ## 0.0.14
 
 ### Patch Changes